+++ packages/terra-card/tests/jest/Card.test.jsx
@@ -6,44 +6,44 @@ describe('Card', () => {
 
   // Snapshot Tests
   it('should render a default component', () => {
-    const wrapper = shallow(defaultRender);
+    const wrapper = enzyme.shallow(defaultRender);
     expect(wrapper).toMatchSnapshot();
   });
 
   // Prop Tests
   it('should use the default value when no value is given', () => {
-    const wrapper = shallow(defaultRender);
+    const wrapper = enzyme.shallow(defaultRender);
     expect(wrapper.find('.card'));
     expect(wrapper.find('.default'));
   });
 
   it('should use the default variant when prop is provided just like default', () => {
-    const wrapper = shallow(<Card variant="default" />);
+    const wrapper = enzyme.shallow(<Card variant="default" />);
     expect(wrapper.find('.default'));
     expect(wrapper).toMatchSnapshot();
   });
 
   it('should use the raised variant when prop is provided', () => {
-    const wrapper = shallow(<Card variant="raised" />);
+    const wrapper = enzyme.shallow(<Card variant="raised" />);
     expect(wrapper.find('.raised'));
     expect(wrapper).toMatchSnapshot();
   });
 
   // Structure Tests
   it('should have the class card', () => {
-    const wrapper = shallow(defaultRender);
+    const wrapper = enzyme.shallow(defaultRender);
     expect(wrapper.prop('className')).toContain('card');
   });
 
   // Custom Prop Test
   it('it should pass in a custom prop', () => {
-    const wrapper = shallow(<Card id="testCard" />);
+    const wrapper = enzyme.shallow(<Card id="testCard" />);
     expect(wrapper.props().id).toEqual('testCard');
   });
 
   // Visually hidden text
   it('it should pass in visually hidden text', () => {
-    const wrapper = shallow(<Card visuallyHiddenText="This is screen reader only text" />);
+    const wrapper = enzyme.shallow(<Card visuallyHiddenText="This is screen reader only text" />);
     expect(wrapper).toMatchSnapshot();
   });
 
@@ -52,7 +52,7 @@ describe('Card', () => {
       .mockReturnValue({
         className: 'orion-fusion-theme',
       });
-    const wrapper = shallow(defaultRender);
+    const wrapper = enzyme.shallow(defaultRender);
     expect(wrapper).toMatchSnapshot();
   });
 });