+++ packages/terra-core-docs/src/terra-dev-site/doc/search-field/Examples.2/SearchFieldDisableAutoFormatAttrs.jsx
@@ -3,6 +3,8 @@ import SearchField from 'terra-search-field';
 
 const SearchFieldDisableAutoFormatAttrs = () => (
   <SearchField
+    isLabelVisible
+    groupName="Find a clinic"
     inputAttributes={{
       autoCorrect: 'off',
       autoComplete: 'off',