+++ packages/terra-demographics-banner/tests/jest/DemographicsBannerDisplay.test.jsx
@@ -3,12 +3,12 @@ import Image from 'terra-image';
 import DemographicsBannerDisplay from '../../src/DemographicsBannerDisplay';
 
 it('renders a blank banner properly', () => {
-  const banner = shallow(<DemographicsBannerDisplay />);
+  const banner = enzyme.shallow(<DemographicsBannerDisplay />);
   expect(banner).toMatchSnapshot();
 });
 
 it('renders a banner that has basic information', () => {
-  const banner = shallow(<DemographicsBannerDisplay
+  const banner = enzyme.shallow(<DemographicsBannerDisplay
     age="24"
     dateOfBirth="March 30, 2017"
     gender="F"
@@ -19,7 +19,7 @@ it('renders a banner that has basic information', () => {
 });
 
 it('renders the banner that contains additional information', () => {
-  const banner = shallow(<DemographicsBannerDisplay
+  const banner = enzyme.shallow(<DemographicsBannerDisplay
     applicationContent={<span className="risk-score">5%</span>}
     age="25 Years"
     applicationRows={[{ Address: '1501 Walnut St. Apt #123', City: 'Kansas City MO' }, { ZIP: 64108 }]}
@@ -35,7 +35,7 @@ it('renders the banner that contains additional information', () => {
 });
 
 it('renders the banner appropriately for a deceased person', () => {
-  const banner = shallow(<DemographicsBannerDisplay
+  const banner = enzyme.shallow(<DemographicsBannerDisplay
     age="24"
     gender="F"
     dateOfBirth="June 15, 2014"
@@ -47,7 +47,7 @@ it('renders the banner appropriately for a deceased person', () => {
 });
 
 it('renders the banner appropriately for a person with gestational and post menstrual dates', () => {
-  const banner = shallow(<DemographicsBannerDisplay
+  const banner = enzyme.shallow(<DemographicsBannerDisplay
     applicationContent={<span className="risk-score">5%</span>}
     age="25 Years"
     applicationRows={[{ Address: '1501 Walnut St. Apt #123', City: 'Kansas City MO' }, { ZIP: 64108 }]}
@@ -67,17 +67,17 @@ it('renders the banner appropriately for a person with gestational and post mens
 });
 
 it('renders gestational date with a default label when the label is not provided', () => {
-  const banner = shallow(<DemographicsBannerDisplay gestationalAge="April 5, 2016" />);
+  const banner = enzyme.shallow(<DemographicsBannerDisplay gestationalAge="April 5, 2016" />);
   expect(banner).toMatchSnapshot();
 });
 
 it('renders post menstrual date with a default label when the label is not provided', () => {
-  const banner = shallow(<DemographicsBannerDisplay postMenstrualAge="April 5, 2016" />);
+  const banner = enzyme.shallow(<DemographicsBannerDisplay postMenstrualAge="April 5, 2016" />);
   expect(banner).toMatchSnapshot();
 });
 
 it('renders the banner properly for a deceased person with additional application content', () => {
-  const banner = shallow(<DemographicsBannerDisplay
+  const banner = enzyme.shallow(<DemographicsBannerDisplay
     applicationContent={<span className="risk-score">5%</span>}
     age="25 Years"
     applicationRows={[{ Address: '1501 Walnut St. Apt #123', City: 'Kansas City MO' }, { ZIP: 64108 }]}