+++ packages/terra-form-select/tests/jest/ComboboxField.test.jsx
@@ -1,6 +1,4 @@
 import React from 'react';
-/* eslint-disable-next-line import/no-extraneous-dependencies */
-import { shallowWithIntl } from 'terra-enzyme-intl';
 import ComboboxField from '../../src/ComboboxField';
 
 it('should render ComboboxField component with max height', () => {
@@ -14,6 +12,6 @@ it('should render ComboboxField component with max height', () => {
     </ComboboxField>
   );
 
-  expect(shallowWithIntl(select).dive()).toMatchSnapshot();
+  expect(enzymeIntl.shallowWithIntl(select).dive()).toMatchSnapshot();
 });