+++ packages/terra-form-select/tests/jest/Option.test.jsx
@@ -6,12 +6,12 @@ import Option from '../../src/shared/_Option';
 
 describe('Option', () => {
   it('should render a default Option', () => {
-    const wrapper = shallow(<Option value="value" display="display" />);
+    const wrapper = enzyme.shallow(<Option value="value" display="display" />);
     expect(wrapper).toMatchSnapshot();
   });
 
   it('should render a disabled Option', () => {
-    const wrapper = shallow(<Option disabled value="value" display="display" />);
+    const wrapper = enzyme.shallow(<Option disabled value="value" display="display" />);
     expect(wrapper).toMatchSnapshot();
   });
 
@@ -21,7 +21,7 @@ describe('Option', () => {
   });
 
   it('correctly applies the theme context className', () => {
-    const wrapper = mount(
+    const wrapper = enzyme.mount(
       <ThemeContextProvider theme={{ className: 'orion-fusion-theme' }}>
         <Option value="value" display="display" />
       </ThemeContextProvider>,