+++ packages/terra-search-field/tests/jest/SearchField.test.jsx
@@ -1,42 +1,40 @@
 import React from 'react';
 import ThemeContextProvider from 'terra-theme-context/lib/ThemeContextProvider';
 
-/* eslint-disable-next-line import/no-extraneous-dependencies */
-import { shallowWithIntl, mountWithIntl } from 'terra-enzyme-intl';
 import SearchField from '../../src/SearchField';
 
 describe('Snapshots', () => {
   it('renders a basic search field', () => {
-    const searchField = shallowWithIntl(<SearchField />).dive();
+    const searchField = enzymeIntl.shallowWithIntl(<SearchField />).dive();
     expect(searchField).toMatchSnapshot();
   });
 
   it('renders a search field with a label', () => {
-    const searchField = shallowWithIntl(<SearchField groupName="TestGroup" isLabelVisible />).dive();
+    const searchField = enzymeIntl.shallowWithIntl(<SearchField groupName="TestGroup" isLabelVisible />).dive();
     expect(searchField.find('label').text()).toEqual('TestGroup');
     expect(searchField).toMatchSnapshot();
   });
 
   it('renders a search field with a placeholder', () => {
-    const searchField = shallowWithIntl(<SearchField placeholder="Test" />).dive();
+    const searchField = enzymeIntl.shallowWithIntl(<SearchField placeholder="Test" />).dive();
     expect(searchField.find('input').prop('placeholder')).toEqual('Test');
     expect(searchField).toMatchSnapshot();
   });
 
   it('renders a search field with a value', () => {
-    const searchField = shallowWithIntl(<SearchField value="Test" />).dive();
+    const searchField = enzymeIntl.shallowWithIntl(<SearchField value="Test" />).dive();
     expect(searchField.find('input').prop('value')).toEqual('Test');
     expect(searchField).toMatchSnapshot();
   });
 
   it('renders a search field with a defaulted value', () => {
-    const searchField = shallowWithIntl(<SearchField defaultValue="Default" />).dive();
+    const searchField = enzymeIntl.shallowWithIntl(<SearchField defaultValue="Default" />).dive();
     expect(searchField.find('input').prop('defaultValue')).toEqual('Default');
     expect(searchField).toMatchSnapshot();
   });
 
   it('renders a disabled search field with a value', () => {
-    const searchField = shallowWithIntl(<SearchField isDisabled />).dive();
+    const searchField = enzymeIntl.shallowWithIntl(<SearchField isDisabled />).dive();
     searchField.instance().updateSearchText('Test');
     expect(searchField.find('input').prop('disabled')).toBe(true);
     expect(searchField.find('Button').prop('isDisabled')).toBe(true);
@@ -44,7 +42,7 @@ describe('Snapshots', () => {
   });
 
   it('renders a search field that displays as a block to fill its container', () => {
-    const searchField = shallowWithIntl(<SearchField isBlock />).dive();
+    const searchField = enzymeIntl.shallowWithIntl(<SearchField isBlock />).dive();
     searchField.instance().updateSearchText('Test');
     expect(searchField.find('.search-field').prop('className')).toContain('block');
     expect(searchField).toMatchSnapshot();
@@ -52,31 +50,31 @@ describe('Snapshots', () => {
 
   it('passes in inputRefCallback as the refCallback prop of the Input component', () => {
     const inputRefCallback = jest.fn();
-    const searchField = mountWithIntl(<SearchField inputRefCallback={inputRefCallback} />).children();
+    const searchField = enzymeIntl.mountWithIntl(<SearchField inputRefCallback={inputRefCallback} />).children();
     expect(inputRefCallback).toBeCalled();
     expect(searchField).toMatchSnapshot();
   });
 
   it('renders a search field with an aria-label', () => {
-    const searchField = shallowWithIntl(<SearchField inputAttributes={{ 'aria-label': 'Search Field' }} />).dive();
+    const searchField = enzymeIntl.shallowWithIntl(<SearchField inputAttributes={{ 'aria-label': 'Search Field' }} />).dive();
     expect(searchField.find('input').prop('aria-label')).toEqual('Search Field');
     expect(searchField).toMatchSnapshot();
   });
 
   it('renders a search field with an aria-label using prop', () => {
-    const searchField = shallowWithIntl(<SearchField inputAttributes={{ ariaLabel: 'Search Field' }} />).dive();
-    expect(searchField.find('input').prop('aria-label')).toEqual('Terra.searchField.search');
+    const searchField = enzymeIntl.shallowWithIntl(<SearchField inputAttributes={{ ariaLabel: 'Search Field' }} />).dive();
+    expect(searchField.find('input').prop('aria-label')).toEqual('');
     expect(searchField).toMatchSnapshot();
   });
 
   it('renders a search field such that custom styles are applied', () => {
-    const searchField = shallowWithIntl(<SearchField inputAttributes={{ className: 'test-class' }} />).dive();
+    const searchField = enzymeIntl.shallowWithIntl(<SearchField inputAttributes={{ className: 'test-class' }} />).dive();
     expect(searchField.find('input').prop('className')).toContain('test-class');
     expect(searchField).toMatchSnapshot();
   });
 
   it('clears form with clear method', () => {
-    const searchField = shallowWithIntl(<SearchField />).dive();
+    const searchField = enzymeIntl.shallowWithIntl(<SearchField />).dive();
     searchField.instance().updateSearchText('amp');
 
     expect(searchField.instance().searchText).toBe('amp');
@@ -85,7 +83,7 @@ describe('Snapshots', () => {
   });
 
   it('correctly applies the theme context className', () => {
-    const searchField = mountWithIntl(
+    const searchField = enzymeIntl.mountWithIntl(
       <ThemeContextProvider theme={{ className: 'orion-fusion-theme' }}>
         <SearchField />
       </ThemeContextProvider>,
@@ -97,7 +95,7 @@ describe('Snapshots', () => {
 describe('Manual Search', () => {
   it('triggers search on button click', () => {
     const onSearch = jest.fn();
-    const searchField = shallowWithIntl(<SearchField onSearch={onSearch} />).dive();
+    const searchField = enzymeIntl.shallowWithIntl(<SearchField onSearch={onSearch} />).dive();
     searchField.instance().updateSearchText('Te');
     expect(onSearch).not.toBeCalled();
     searchField.find('.search-field').childAt(1).simulate('click');
@@ -111,7 +109,7 @@ describe('Manual Search', () => {
     document.body.appendChild(container);
 
     const onSearch = jest.fn();
-    const searchField = mountWithIntl(<SearchField onSearch={onSearch} />, {
+    const searchField = enzymeIntl.mountWithIntl(<SearchField onSearch={onSearch} />, {
       attachTo: document.querySelector('#container'),
     });
 
@@ -126,7 +124,7 @@ describe('Manual Search', () => {
 
   it('does not trigger search if default minimum search text has not been met', () => {
     const onSearch = jest.fn();
-    const searchField = shallowWithIntl(<SearchField onSearch={onSearch} />).dive();
+    const searchField = enzymeIntl.shallowWithIntl(<SearchField onSearch={onSearch} />).dive();
     searchField.setState({ searchText: 'T' });
 
     expect(onSearch).not.toBeCalled();
@@ -136,7 +134,7 @@ describe('Manual Search', () => {
 
   it('does not trigger search if minimum search text has not been met', () => {
     const onSearch = jest.fn();
-    const searchField = shallowWithIntl(<SearchField onSearch={onSearch} minimumSearchTextLength={5} />).dive();
+    const searchField = enzymeIntl.shallowWithIntl(<SearchField onSearch={onSearch} minimumSearchTextLength={5} />).dive();
     searchField.setState({ searchText: 'Sear' });
 
     expect(onSearch).not.toBeCalled();
@@ -145,7 +143,7 @@ describe('Manual Search', () => {
   });
 
   it('does not search when callback is not provided', () => {
-    const searchField = shallowWithIntl(<SearchField minimumSearchTextLength={5} />).dive();
+    const searchField = enzymeIntl.shallowWithIntl(<SearchField minimumSearchTextLength={5} />).dive();
     searchField.setState({ searchText: 'Searc' });
 
     searchField.find('.search-field').childAt(1).simulate('click'); // Verifies we do not attempt to call an undefined function.
@@ -164,7 +162,7 @@ describe('Auto Search', () => {
 
   it('triggers search after delay on text change', () => {
     const onSearch = jest.fn();
-    const searchField = shallowWithIntl(<SearchField onSearch={onSearch} />).dive();
+    const searchField = enzymeIntl.shallowWithIntl(<SearchField onSearch={onSearch} />).dive();
 
     searchField.find('.input').simulate('change', { target: { value: 'Te' } });
 
@@ -175,7 +173,7 @@ describe('Auto Search', () => {
 
   it('only triggers one search if a manual search is triggered before automatic search', () => {
     const onSearch = jest.fn();
-    const searchField = shallowWithIntl(<SearchField onSearch={onSearch} />).dive();
+    const searchField = enzymeIntl.shallowWithIntl(<SearchField onSearch={onSearch} />).dive();
 
     searchField.find('.input').simulate('change', { target: { value: 'Te' } });
 
@@ -191,7 +189,7 @@ describe('Auto Search', () => {
   it('does not trigger search if minimum text length is not met', () => {
     jest.useFakeTimers();
     const onSearch = jest.fn();
-    const searchField = shallowWithIntl(<SearchField onSearch={onSearch} minimumSearchTextLength={5} />).dive();
+    const searchField = enzymeIntl.shallowWithIntl(<SearchField onSearch={onSearch} minimumSearchTextLength={5} />).dive();
 
     searchField.find('.input').simulate('change', { target: { value: 'Sear' } });
 
@@ -204,7 +202,7 @@ describe('Auto Search', () => {
     jest.useFakeTimers();
     const onSearch = jest.fn();
     const onInvalidSearch = jest.fn();
-    const searchField = shallowWithIntl(<SearchField onSearch={onSearch} onInvalidSearch={onInvalidSearch} minimumSearchTextLength={5} />).dive();
+    const searchField = enzymeIntl.shallowWithIntl(<SearchField onSearch={onSearch} onInvalidSearch={onInvalidSearch} minimumSearchTextLength={5} />).dive();
 
     searchField.find('.input').simulate('change', { target: { value: 'Sear' } });
 
@@ -216,7 +214,7 @@ describe('Auto Search', () => {
   it('uses standard timeout for search delay when not provided', () => {
     jest.useFakeTimers();
     jest.spyOn(global, 'setTimeout');
-    const searchField = shallowWithIntl(<SearchField />).dive();
+    const searchField = enzymeIntl.shallowWithIntl(<SearchField />).dive();
     searchField.find('.input').simulate('change', { target: {} });
     expect(setTimeout).toBeCalledWith(expect.anything(), 2500);
   });
@@ -225,7 +223,7 @@ describe('Auto Search', () => {
     jest.useFakeTimers();
     jest.spyOn(global, 'setTimeout');
     const onSearch = jest.fn();
-    const searchField = shallowWithIntl(<SearchField searchDelay={1000} onSearch={onSearch} />).dive();
+    const searchField = enzymeIntl.shallowWithIntl(<SearchField searchDelay={1000} onSearch={onSearch} />).dive();
 
     searchField.find('.input').simulate('change', { target: {} });
     jest.advanceTimersByTime(2500);
@@ -242,7 +240,7 @@ describe('Auto Search', () => {
 
   it('should call onChange when button is selected', () => {
     const onChange = jest.fn();
-    const searchField = shallowWithIntl(<SearchField onChange={onChange} />).dive();
+    const searchField = enzymeIntl.shallowWithIntl(<SearchField onChange={onChange} />).dive();
 
     searchField.find('.input').simulate('change', { target: {} });
     expect(onChange).toBeCalled();