+++ packages/terra-core-docs/CHANGELOG.md
@@ -4,11 +4,13 @@
 
 * Added
   * Added a test example to validate focus when none of the buttons are selected.
+  * Added visual labels for `terra-search-field` examples. 
 
 * Updated
   * Updated Draggable list item example for selection state in `terra-list`.
   * Updated `terra-status` documentation with deprecation notice and badge.
   * Updated `terra-html-table` documentation with deprecation notice and badge.
+  * Updated test and doc examples for `terra-progress-bar`
 
 ## 1.56.0 - (January 9, 2024)